# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MSM Alimentos — API pública",
    "version": "1.2.0",
    "description": "Endpoints públicos do site institucional MSM. Para integradores (MullerBot, monitoramento, automações). Responses incluem header `X-API-Version` (`1.2`). CORS aberto em endpoints públicos (`*`); endpoints com auth Bearer não devem ser chamados de browser sem origin allowlist explícita.",
    "contact": { "name": "MSM Alimentos", "email": "comercial@msm.com.br" }
  },
  "servers": [
    { "url": "https://site-msm-production.up.railway.app", "description": "Produção (temporário Railway)" },
    { "url": "http://localhost:3000", "description": "Local dev" }
  ],
...
  "paths": {
...
    "/api/admin/audit-stats": {
      "get": {
        "tags": ["internal"],
        "summary": "Agregações de admin_activity_log",
        "description": "Retorna estatísticas agregadas (top actions/tables/users + série diária) do audit log. Diferente do /audit-trail que retorna linhas brutas. Útil para dashboard.",
        "parameters": [
          { "name": "from", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          { "name": "to", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          { "name": "top", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 50, "default": 10 } }
        ],
        "responses": {
          "200": {
            "description": "Agregações",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "total": { "type": "integer" },
                    "by_action": { "type": "array" },
                    "by_table": { "type": "array" },
                    "by_user": { "type": "array" },
                    "per_day": { "type": "array" }
                  }
                }
              }
            }
          },
          "400": { "description": "Query inválida" },
          "401": { "description": "Não autenticado como admin" }
        }
      }
    },
...
    "/api/admin/audit-trail": {
      "get": {
        "tags": ["internal"],
        "summary": "Export filtrado de admin_activity_log",
        "description": "Retorna registros de auditoria com filtros (date range, user, action, table). Auth via session cookie + admin_profiles (não Bearer). Suporta paginação (limit/offset, max 1000) e formatos JSON ou CSV (RFC 4180).",
        "parameters": [
          { "name": "from", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          { "name": "to", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          { "name": "user_email", "in": "query", "schema": { "type": "string", "format": "email" } },
          { "name": "user_id", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "action", "in": "query", "schema": { "type": "string", "maxLength": 40 } },
          { "name": "table_name", "in": "query", "schema": { "type": "string", "maxLength": 40 } },
          { "name": "record_id", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 1000, "default": 100 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "minimum": 0, "default": 0 } },
          { "name": "format", "in": "query", "schema": { "type": "string", "enum": ["json", "csv"], "default": "json" } }
        ],
        "responses": {
          "200": {
            "description": "Registros (JSON ou CSV)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "rows": { "type": "array" },
                    "count": { "type": "integer", "description": "Total matching antes da paginação" },
                    "has_more": { "type": "boolean" },
                    "query": { "type": "object" }
                  }
                }
              },
              "text/csv": { "schema": { "type": "string" } }
            }
          },
          "400": { "description": "Filtros inválidos (validação Zod)" },
          "401": { "description": "Não autenticado como admin" }
        }
      }
    },
...
    "/api/csp-report": {
      "post": {
        "tags": ["public"],
        "summary": "Receptor de violações CSP",
        "description": "Chamado pelo browser quando uma diretiva CSP é violada. Aceita formato CSP Level 2 (application/csp-report) e Level 3 (Reporting API). Loga via log.warn e incrementa contador msm_csp_violations_total{directive}. Retorna 204 silenciosamente.",
        "requestBody": {
          "required": true,
          "content": {
            "application/csp-report": {
              "schema": {
                "type": "object",
                "properties": {
                  "csp-report": {
                    "type": "object",
                    "properties": {
                      "document-uri": { "type": "string" },
                      "violated-directive": { "type": "string" },
                      "effective-directive": { "type": "string" },
                      "blocked-uri": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": { "description": "Violação registrada (silenciosa, conforme spec)" },
          "400": { "description": "JSON inválido" }
        }
      }
    },
...
    "/api/cron/audit-archive": {
      "get": {
        "tags": ["internal"],
        "summary": "Archive admin_activity_log antigo (daily)",
        "description": "Apaga registros > AUDIT_RETENTION_DAYS (default 90). Batch máximo 1000 por execução para evitar timeouts. Suporta dry-run via ?dry-run=true.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "dry-run",
            "in": "query",
            "required": false,
            "schema": { "type": "boolean" },
            "description": "Se true, conta quantos seriam apagados sem deletar"
          }
        ],
        "responses": {
          "200": {
            "description": "Archive concluído",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "cutoff_date": { "type": "string", "format": "date-time" },
                    "retention_days": { "type": "integer" },
                    "deleted_count": { "type": "integer" },
                    "remaining_count": { "type": "integer", "nullable": true },
                    "duration_ms": { "type": "integer" }
                  }
                }
              }
            }
          },
          "401": { "description": "CRON_SECRET inválido" },
          "500": { "description": "CRON_SECRET não configurado em produção" }
        }
      }
    },
...
    "/api/cron/process-webhook-queue": {
      "get": {
        "tags": ["internal"],
        "summary": "Processa fila webhook (cron)",
        "description": "Drena até 50 itens da fila pending/failed. Em prod, sem CRON_SECRET configurado retorna 500.",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Processado" },
          "401": { "description": "CRON_SECRET inválido" },
          "500": { "description": "CRON_SECRET não configurado em produção" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Header `Authorization: Bearer <SECRET>` (REVALIDATE_SECRET ou CRON_SECRET conforme rota)"
      }
    },
...
  }
}